Built framework
Built a basic framework including a SceneManager script, Scene Manager game object,  Sound Manager script and game object, 6 scene's "end" "intro" and levels 1-4. I figure this would be a good skeleton we can start to fill in with content. oh ya and i put a buttion on the top left side of the screne that allows you to progress to the next scene by clicking it and what scene your in in green text next to it.
the scene manager allows you to either load a scene by name or progress through the "Scene In Build" index in the build setting window. also there is a sound manager tht allows music to play constant though out the entire game.